requirements.txt
blinker==1.4
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
filelock==3.15.4
Flask==2.2.5
Flask-Cors==4.0.1
fonttools==4.53.1
fsspec==2024.6.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.7.1
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
opencv-python==4.10.0.84
packaging==24.1
pandas==2.1.4
pillow==10.4.0
psutil==5.9.5
py-cpuinfo==9.0.0
PyMySQL==1.1.1
pyparsing==3.1.2
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
scipy==1.13.1
seaborn==0.13.1
setuptools==73.0.1
six==1.16.0
sympy==1.13.2
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.2.82
ultralytics-thop==2.0.5
urllib3==2.0.7
Werkzeug==3.0.3
